requirements.txt
aiohttp==3.8.1
asyncio==3.4.3
beautifulsoup4==4.12.2
mysql-connector-python==8.0.33
requests==2.31.0
urllib3==1.26.15
